refactor: split morphology lexicon blob into module directory
Extracted massive `static LEXICON` into `data.rs` and DTOs into `model.rs` while keeping lookup functions and public API in `mod.rs`. This improves encapsulation and reduces file size.
