+++ .jules/atlas.md
@@ -20,3 +20,11 @@
 **Tangle:** Several modules under `src/tools/` (specifically `cache`, `report`, and `ui`) and `src/semantic/assembly/` (`model`) were exposed as `pub mod`, breaking encapsulation by exposing internal implementation details to the public API.
 **Blueprint:** Modified `src/tools/mod.rs` and `src/semantic/assembly/mod.rs` to restrict these modules with `pub(crate) mod`.
 **Stability:** Achieved higher cohesion by keeping the public API surface minimal and ensuring internal structures don't leak out of their intended domains.
+
+## [Splitting The Blob: Lexicon]
+**Tangle:** `src/morphology/lexicon.rs` was a monolithic file (~2800 lines) mixing DTOs (`LexiconEntry`, operator enums), lookup logic, and a massive static data map (`LEXICON`).
+**Blueprint:**
+1. Created `src/morphology/lexicon/` module.
+2. Extracted DTOs to `src/morphology/lexicon/model.rs`.
+3. Extracted massive static data to `src/morphology/lexicon/data.rs`.
+4. Moved logic and re-exports to `src/morphology/lexicon/mod.rs`.
